+++ library/core/common/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/core/common/utils/constants/logging/LogTags.kt
@@ -18,14 +18,10 @@
 package com.mihaicristiancondrea.android.libs.apptoolkit.core.common.utils.constants.logging
 
 const val STARTUP_LOG_TAG: String = "StartupActivity"
-const val DISPLAY_SETTINGS_LOG_TAG: String = "DisplaySettings"
 const val CLIPBOARD_HELPER_LOG_TAG: String = "ClipboardHelper"
 const val CONSENT_LOG_TAG: String = "Consent"
 const val GENERAL_SETTINGS_LOG_TAG: String = "GeneralSettingsActivity"
 const val FCM_LOG_TAG: String = "FirebaseNotifications"
-const val FAQ_LOG_TAG: String = "FaqRepository"
 const val THEME_SETTINGS_LOG_TAG: String = "ThemeSettingsList"
 const val ABOUT_SETTINGS_LOG_TAG: String = "AboutSettings"
 const val INFO_MESSAGE_LOG_TAG: String = "InfoMessageSection"
-const val SELECT_STARTUP_DIALOG_LOG_TAG: String = "SelectStartupDialog"
-const val SELECT_LANGUAGE_DIALOG_LOG_TAG: String = "SelectLanguageDialog"